+++ geopandas/geodataframe.py
@@ -1242,7 +1242,7 @@ properties': {'col1': 'name1'}, 'geometry': {'type': 'Point', 'coordinates': (1.
         """
         from geopandas.io.geoarrow import ArrowTable, geopandas_to_arrow
 
-        table = geopandas_to_arrow(
+        table, _ = geopandas_to_arrow(
             self,
             index=index,
             geometry_encoding=geometry_encoding,
@@ -1256,8 +1256,9 @@ properties': {'col1': 'name1'}, 'geometry': {'type': 'Point', 'coordinates': (1.
         path,
         index=None,
         compression="snappy",
-        schema_version=None,
+        geometry_encoding="WKB",
         write_covering_bbox=False,
+        schema_version=None,
         **kwargs,
     ):
         """Write a GeoDataFrame to the Parquet format.
@@ -1279,6 +1280,10 @@ properties': {'col1': 'name1'}, 'geometry': {'type': 'Point', 'coordinates': (1.
             output except `RangeIndex` which is stored as metadata only.
         compression : {'snappy', 'gzip', 'brotli', None}, default 'snappy'
             Name of the compression to use. Use ``None`` for no compression.
+        geometry_encoding : {'WKB', 'geoarrow'}, default 'WKB'
+            The encoding to use for the geometry columns. Defaults to "WKB"
+            for maximum interoperability. Specify "geoarrow" to use one of the
+            native GeoArrow-based single-geometry type encodings.
         schema_version : {'0.1.0', '0.4.0', '1.0.0', None}
             GeoParquet specification version; if not provided will default to
             latest supported version.
@@ -1317,6 +1322,7 @@ properties': {'col1': 'name1'}, 'geometry': {'type': 'Point', 'coordinates': (1.
             self,
             path,
             compression=compression,
+            geometry_encoding=geometry_encoding,
             index=index,
             schema_version=schema_version,
             write_covering_bbox=write_covering_bbox,